Cannot find a valid compressor lz4
I've updated common.sh to include lz4 support when building erofs-utils. This should resolve the "Cannot find a valid compressor lz4" error. The user will need to run the script again, which will trigger a rebuild of erofs-utils with the new configuration. After that, the repacking process should work as expected.